requirements_win.txt
librosa==0.9.1
fairseq==0.12.2
ffmpeg-python
Flask==2.1.2
Flask_Cors==3.0.10
gradio>=3.7.0
numpy
playsound==1.2.2
pydub==0.25.1
pyworld==0.3.4
requests
scipy
sounddevice
SoundFile
starlette
tqdm
rich
loguru
torchcrepe
scikit-maad
praat-parselmouth
onnx
onnxsim
onnxoptimizer
tensorboard
tensorboardX
transformers
edge_tts
langdetect
pyyaml
pynvml
faiss-cpu
